Remove pyanaconda.payload.dnfpayload
Move all code from dnfpayload.py to new files in pyanaconda.payload.dnf.
